+++ resotocore/tests/resotocore/util_test.py
@@ -4,6 +4,8 @@ import shutil
 import pytest
 from aiostream import stream
 from copy import deepcopy
+from dataclasses import dataclass
+from typing import Any
 
 from resotocore.util import (
     AccessJson,
@@ -16,6 +18,7 @@ from resotocore.util import (
     del_value_in_path,
     deep_merge,
     partition_by,
+    merge_json_elements,
 )
 
 
@@ -120,3 +123,54 @@ def test_deep_merge() -> None:
     l = {"a": {"b": 1, "d": 2}, "d": 2, "e": 4}
     r = {"a": {"c": 1, "d": 3}, "d": 1}
     assert deep_merge(l, r) == {"a": {"b": 1, "c": 1, "d": 3}, "d": 1, "e": 4}
+
+
+def test_merge_values1() -> None:
+    a = {"a": {"foo": {"first": "first", "last": "laaaast"}}, "b": {"bar": 123}, "c": [6, 7]}
+    b = {"a": {"foo": {"last": "last"}}, "b": {"baz": 456}, "c": [8, 9]}
+    assert deep_merge(a, b) == {
+        "a": {"foo": {"first": "first", "last": "last"}},
+        "b": {"bar": 123, "baz": 456},
+        "c": [8, 9],
+    }
+
+
+def test_merge_values() -> None:
+    a = {"a": {"foo": {"first": "first", "last": "laaaast"}}, "b": {"bar": 123}, "c": [6, 7]}
+    b = {"a": {"foo": {"last": "last"}}, "b": {"baz": 456}, "c": [8, 9]}
+    assert merge_json_elements(a, b) == {
+        "a": {"foo": {"first": "first", "last": "last"}},
+        "b": {"bar": 123, "baz": 456},
+        "c": [8, 9],
+    }
+
+    def concat_lists(a: Any, b: Any) -> Any:
+        if isinstance(a, list) and isinstance(b, list):
+            return a + b
+        return b
+
+    assert merge_json_elements([1], [2], merge_strategy=concat_lists) == [1, 2]
+
+    def add_comments(a: Any, b: Any) -> Any:
+        if isinstance(a, (str, int)):
+            return f"{a} replaced by {b}"
+        return b
+
+    assert merge_json_elements({"a": "foo", "b": "baz"}, {"a": "bar", "c": "c_val"}, merge_strategy=add_comments) == {
+        "a": "foo replaced by bar",
+        "b": "baz",
+        "c": "c_val",
+    }
+
+    @dataclass
+    class OverriddenValue:
+        existing: Any
+        override: Any
+
+    def wrap_in_class(a: Any, b: Any) -> Any:
+        return OverriddenValue(a, b)
+
+    assert merge_json_elements({"a": "foo", "b": {"c": "c_val"}}, {"a": "bar", "b": "foo"}, wrap_in_class) == {
+        "a": OverriddenValue("foo", "bar"),
+        "b": OverriddenValue({"c": "c_val"}, "foo"),
+    }
